cut out ping provider related classes as a file
